# api-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de BodyTech",
        "description": "Documentación de la API de BodyTech con Swagger",
        "contact": {
            "email": "soporte@bodytech.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Obtener listado de actividades",
                "description": "Listar actividades con filtros opcionales.",
                "operationId": "53e7a0b67f6097ed90245000764b2e5e",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filtrar por ID de usuario",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "description": "Filtrar por acción (ej. login, logout)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Fecha de inicio (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Fecha de fin (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de actividades paginado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Registrar una nueva actividad",
                "description": "Registrar una nueva actividad.",
                "operationId": "8aa12114961e22767ecdad7c24be367e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "action"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "action": {
                                        "type": "string",
                                        "example": "login"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Usuario inició sesión"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Actividad creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/activities/{id}": {
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Actualizar una actividad",
                "description": "Actualizar una actividad.",
                "operationId": "064379988a5fe2c974ae5ab18f0cddfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la actividad a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "action"
                                ],
                                "properties": {
                                    "action": {
                                        "type": "string",
                                        "example": "logout"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Usuario cerró sesión"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Eliminar una actividad",
                "description": "Eliminar una actividad.",
                "operationId": "d822403ad9a85feb951b1073858a072d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la actividad a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Actividad eliminada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Actividad",
                "description": "Modelo de actividad del usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "action": {
                        "type": "string",
                        "example": "login"
                    },
                    "description": {
                        "type": "string",
                        "example": "El usuario inició sesión"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-28T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-02-28T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}